Added a winScreen() method and a stopGameTimer() method and a new gameOver boolean.
WinScreen()
Opens a new window with an animated image and a panel that says
"Winner Winner Java Dinner"
It also sets gameOver to true which now disables the timer and the ability to move
tiles.

stopGameTimer()
stops the timer so that when you've won the game it doesnt keep on counting.